+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * server startup output with image version
   * readiness probe for stool server
 * update to junit 5.7.0
+* renamed environment to values
 
 
 ### 6.0.0 (2020-10-12)